make DISTINCT ON plans work again (WIP)
Insert RemoteSubplan paths at appropriate places, to generate plans
like this:

                                          QUERY PLAN
    --------------------------------------------------------------------------------------
     HashAggregate  (cost=153.86..155.86 rows=200 width=4)
       Group Key: b
       ->  Remote Subquery Scan on all (dn1,dn2)  (cost=100.00..148.76 rows=2040 width=4)
             ->  Seq Scan on t1  (cost=0.00..30.40 rows=2040 width=4)

or

                                          QUERY PLAN
    --------------------------------------------------------------------------------------
     Unique  (cost=242.54..262.94 rows=200 width=4)
       ->  Remote Subquery Scan on all (dn1,dn2)  (cost=242.54..257.84 rows=2040 width=4)
             ->  Sort  (cost=142.54..147.64 rows=2040 width=4)
                   Sort Key: b
                   ->  Seq Scan on t1  (cost=0.00..30.40 rows=2040 width=4)

These changes may not be entirely correct yet - the "Unique" one seems
to be missing a "Sort" node on top of "Remote Subquery Scan", although
that may not be necessary for some distributions (matching the grouping
column).

For the "HashAggregate" plans, it might be possible to use the 2-phase
aggregation, pushing some of the computation to remote nodes (reducing
the amount of data transferred).