generate distributed grouping paths with a combine phase
Generate grouping paths with an extra aggregate step before the
remote subplan, looking like this:

    Finalize GroupAggregate
        ->  Remote Subquery Scan
            ->  Combine Aggregate
                ->  Gather
                    ->  Partial GroupAggregate
                        -> ...

The idea is that the combine phase may significantly reduce the
number of rows transferred to the coordinators (and processed on
them), possibly up to 1/N where N is the number of workers.

The plans don't seem to be really chosen at this point, most likely
because of incorrect costing that makes them look more expensive
than the paths without the combine step.