skip RenameSequenceGTM() for SET SCHEMA with the same schema
Commit bc4996e6 in upstream changed how CheckSetNamespace() handles
moving to the same namespace, e.g. when executing

    ALTER SEQUENCE s1.s SET SCHEMA s1;

On PostgreSQL 9.5 this fails with an error, but since bc4996e6 the
command is silently ignored (which was already the case for other
object types - ALTER EXTENSION etc).

XL however relied on the ERROR interrupting the program flow before
the call to RenameSequenceGTM(), so after bc4996e6 it was executed
anyway.

GTM however does not expect unnecessary sequence renames, and treats
this as an error in seq_add_seqinfo.