remove 'missing redistribute_path' error from recurse_set_operations
I came to the conclusion that we probably don't need to redistribute
the path at this place. We're producing a generic subquery path with
particular distribution - if if a path with different distribution
is needed, it should be responsibility of the caller to do that.

The assumption is that there might be multiple possible plans, each
requiring a path with a different distribution.