make sure explicit Sorts in Merge Append are pushed to remote nodes
When redistributing subpaths of Merge Append, an explicit sort may be
required. In that case we want to push it to the remote node instead
of performing it locally.

This was originally done in create_remotescan_plan(), i.e. after all
the planning was done. After the upper-planner pathification that's
not very elegant, so this patch moves this logic to pathnode.c.

The design follows what create_merge_append_path was already doing,
we only do the costing in redistribute_path() and the Sort node is
still construted within create_plan().