WIP: mark Sort push-down (through RemoteSubplan) work again correctly
After moving the path redistribution and various bits to pathnode.c,
the explicit Sort nodes needed by Merge Joins with unsorted paths
were not properly pushed to the remote node (but on top of it).

Fixed by a bit of hacking set_joinpath_distribution(). Admittedly,
set_joinpath_distribution() is a mighty beast that would benefit
from a bit of refactoring.

This fixes most of the EXPLAIN regressions, although a bunch of
plans are still slightly different. But that seems to be either
because we managed to find a different (and slightly cheaper)
plan, or due to some other bug (likely in distribution handling).

FIXME This however seems to break portals/transactions regression
tests, resulting in strange crashes in PortalStart due to access
beyond the end of tupDesc->attrs:

    #0  PortalStart (portal=0x124ddf0, params=0x0, eflags=0, snapshot=0x0) at pquery.c:791
    791									InvalidOid :
    Missing separate debuginfos, use: dnf debuginfo-install glibc-2.22-18.fc23.x86_64
    (gdb) bt
    #0  PortalStart (portal=0x124ddf0, params=0x0, eflags=0, snapshot=0x0) at pquery.c:791
    #1  0x00000000008720e5 in exec_bind_message (input_message=0x7ffec76f1790) at postgres.c:2217
    #2  0x000000000087566f in PostgresMain (argc=1, argv=0x122f768, dbname=0x122f408 "regression", username=0x122f3e8 "user") at postgres.c:4788
    #3  0x00000000007e7d63 in BackendRun (port=0x1255ae0) at postmaster.c:4496
    #4  0x00000000007e74c7 in BackendStartup (port=0x1255ae0) at postmaster.c:4170
    #5  0x00000000007e3887 in ServerLoop () at postmaster.c:1806
    #6  0x00000000007e2ebc in PostmasterMain (argc=9, argv=0x122e390) at postmaster.c:1414
    #7  0x00000000007103a0 in main (argc=9, argv=0x122e390) at main.c:228
    (gdb) list
    786
    787    /*
    788     * Set up locator if result distribution is requested
    789     */
    790    keytype = queryDesc->plannedstmt->distributionKey == InvalidAttrNumber ?
    791         InvalidOid :
    792         queryDesc->tupDesc->attrs[queryDesc->plannedstmt->distributionKey-1]->atttypid;
    793    locator = createLocator(
    794         queryDesc->plannedstmt->distributionType,
    795         RELATION_ACCESS_INSERT,
    (gdb) print queryDesc->plannedstmt->distributionKey
    $1 = 3
    (gdb) print queryDesc->tupDesc->natts
    $2 = 2

In transaction.sql, this seems to be triggered by the 'revalidate_bug'
test case starting at like 378.