A very naive way to deal with the fact that RemoteQuery (which EXECUTE DIRECT internally uses) cannot support cursor fetch.
FOR EXECUTE query inside plpgsql uses cursor mechanism to fetch a few tuples at
a time. But that fails badly when the query is a EXECUTE DIRECT statement
because we internally use RemoteQuery to execute that. Instead just fetch 10000
tuples at a time and complain if RemoteQuery returns more than 10000 rows. May
be that's enough for the scenario that we're trying to address where a global
view can be created using EXECUTE DIRECT