minor fixes in 'cluster' regression suite (CREATE TABLE, extra .out files)
The 'CREATE TABLE clstr_tst' was wrong, as it needs to reference the
column and not the foreign key constraint. The table can't have a PK
on 'a' (does not include the distribution key), so remove that from
the expected list of constraints etc.

Also remove the unnecessary cluster_1.out and cluster_2.out files.