fix failures due to distributed update/delete plans in some tests
Some tests are updating multiple columns of input tables, which is
not supported by XL. So instead of failing due to this limitation,
simply replicate the whole table to test whatever was the original
intent of the test.

All the UPDATE/DELETE cases fixed by this commit were failing in
XL 9.5, so none of this should be a regression.