fix failures in 'brin' regression tests
The failures seemed to be caused simply by different parameter
values, but it turned out the tests are actually broken and don't
really work as intended, for two reasons.

Firstly, the script does EXPLAIN on multiple queries and checks if
there's Bitmap Heap Scan etc. but it was only looking at the very
beginning of the result. But XL puts Remote Fast Query Execution
on top of the plan, so these checks were failing (but the WARNING
output was accepted, so no one knew).

Secondly, one of the commands populating the test table was using
subquery with ORDER BY, which is not supported in XL. So the query
was failing, and the results were a bit different from upstream.

So this commit fixes two things:

- Changes the pattern matching to look for the expected operation
anywhere in the plan. Considering that the queries are very simple
(single scan with a single condition), this is unlikely to trigger
false positives or negatives.

- Changes how the 'brintest' table is pupulated, replacing the
single subquery (with ORDER BY) with a few commands producing the
same data set as on upstream.