Push recursive paths to remote nodes when construction the paths
Until now, the pushdown happened in make_recursive_union(), but that
apparently fails to update subroots (and possibly other fields), causing
segfaults in SS_finalize_plan().

Instead move the push-down logic to generate_recursion_path(), which
fixes the issue and is consistent with the overall pathification of
the planner.

With this change XL 9.6 now generates the same plans as XL 9.5.

There remains a lot to be desired, though. Firstly, it fails for queries
with nested recursive unions, because it produces a plan like this:

     ->  Remote Subquery Scan on all (dn1)
           ->  Recursive Union
                 ->  Seq Scan on department
                 ->  CTE Scan on x
                       CTE x
                         ->  Remote Subquery Scan on all (dn1)
                               ->  Recursive Union
                                     ->  Seq Scan on department
                                     ->  Append
                                           ->  WorkTable Scan on q
                                           ->  WorkTable Scan on x

This however fails with 'ERROR: unrecognized node type: 120' error,
because determine_param_types() in execRemote.c does not expect
T_WorkTableScan.

After resolving that issue, it however starts failing with a segfault, as
WorkTableScan nodes try to access the top recursive union node, which is
however executed in a different process, and possibly even on a different
data node (note the "Remote Subquery").

Those issues are however not resolved by this commit, and is left for
the future.