Propagate the distribution up when (root->distribution==NULL).
Without this patch the planner was constricting paths like this:

    -> Material
         -> SubquerySubPath
              -> RemoteSubPath
                   -> BitmapHeapScan

which then led to failures to plan some of the UPDATE and DELETE
queries in updatable_views. This fixes it, and makes the planner
build the same paths as on XL 9.5 (without the RemoteSubPath,
discarding the distribution info).

This resolved all the 'could not plan distributed UPDATE/DELETE'
in updatable_views, but not universally. So either this still is
not entirely correct, or there's another independent bug.