Resolve failures in the cluster regression suite
Distributing tables in this particular test suite does not make
much sense. Firstly, the tables would need to be more complex
to work around 'could not plan distributed update' errors. But
more importantly, we would have ho use ORDER BY to make the
query results stable, which defeats the intent, i.e. testing
that the table is sorted on disk. So just use replicated tables.

Also fix name of the user used in the regression tests.