Resolve failures in guc regression suite related to SAVEPOINTs
The test suite checked GUCs behavior when combined with SAVEPOINTs,
but subtransactions are unsupported in Postgres-XL so this caused
'current transaction is aborted' failues.

Simply remove parts of the test checking how GUCs interact with
subtransactions. Removing the SAVEPOINTs and updating the expected
results would be useless, because that exactly matches already
existing tests in the suite.

There remain several failures in the guc test suite, but those are
independent of SAVEPOINTs, and are more about prepared transactions
vs. access to temporary objects.