Add a user configurable parameter to control the number of rows fetched from the remote side during RemoteSubplan execution.
This allows us to experiment with different sizes more easily. Playing with the
fetch size also exposed couple of problems fixed in this same commit.

1. We were incorrectly forgetting a connection response combiner while
suspending a portal, leading to errors later when we try to buffer the results
because the connection must be used for other queries.

2. The remote cursor name was not getting set properly, thus datanodes
complaining about non-existent cursors.