Resolve failure due to changing cost in xl_join regression suite
The regression failures were caused by slightly fluctuating cost
estimates. In general it's not a good idea to rely on costs being
constant in regression tests, so use EXPLAIN (COSTS OFF) instead.
This stabilizes the output, and resolves the regression failure.

Note: The cost estimates probably change due to stale oldestXmin
reported by one of the nodes, so that analyze does not see some
of the recent changes. That is an independent issue, however, and
it does not really change that relying on cost estimates being
constant is not a good practice.