Resolve failures in foreign_key regression tests
The test case is a new one from upstream, and the failure to distribute
the DELETE (after the ON DELETE rule is created) is already present in
XL 9.5, so accept that as expected behavior.

I've however slightly expanded the test case to also check that after
dropping the rule, the DELETE succeeds. That is expected, as both column
are distribution keys of the two tables.