Accept regression diffs in copydml test case
We don't support triggers yet and hence we accept error messages regarding the
lack of support. Since triggers are not created, the corresponding NOTICEs
raised from the triggers are also not printed.

The core COPY DML functionality is covered by other cases in the test case
hence nothing additional is required.