Make pgbench script names (name prefixes) unique
PostgreSQL 9.6 picks the built-in scripts by prefix, as long as a single
script is matched. This is done even when not using the '-b' option as
in that case pgbench simply uses 'tpcb-like' (and 'pgbench -N' defaults
to 'simple-update').

XL adds two custom scripts that also use the 'bid' column in WHERE
conditions, not just the 'aid' one. Those scripts were however named
'tpcb-like-bid' and 'simple-update-bid' which makes the upstream names
non-unique (when used as prefixes).

This fixes it by simply not keeping the upstream scripts, and replacing
them with the XL version. It might have been possible to massage the
builtin_script[] array to keep both, but it doesn't seem worth it.