Fix warnings about uninitialized vars in pg_dump.c
Three XL-specific fields in getTables() were initialized and used in two
independent if blocks looking like this:

    if (fout->isPostgresXL)
    {
        i_pgxclocatortype = PQfnumber(res, "pgxclocatortype");
        i_pgxcattnum = PQfnumber(res, "pgxcattnum");
        i_pgxc_node_names = PQfnumber(res, "pgxc_node_names");
    }

    if (fout->isPostgresXL)
    {
        ... use the variables ...
    }

Which however confuses the compiler (gcc 5.3.1) which then complains
that the variables are maybe used uninitialized. The fix is simple, just
make the initialization unconditional - if there are no such columns
then PQfnumber() will return -1, but we'll not use the value anyway.