Fix minor issues in the tpcb-like pgbench script
The tpcb-like built-in script in pgbench contained two simple bugs.

It was still using the old \setrandom command to generate the delta
value, instead of the new

    \set delta random(-5000, 5000)

This is mostly an omission in 32d57848458595a487d251b37c2872d86de439ef.

There was also a missing semicolon at the end of one of the commands,
causing cryptic syntax errors.