Disable aggregate paths with extra COMBINE phase
The planner was generating aggregate paths with an additional COMBINE
step pushed to the remote side, like this:

                           QUERY PLAN
    ---------------------------------------------------------
     Finalize Aggregate
       ->  Remote Subquery Scan on all (datanode1,datanode2)
             ->  Partial Aggregate (Combine)
                   ->  Gather
                         ->  Partial Aggregate
                               ->  Parallel Seq Scan on public.t

This was done with the goal to reduce the amount of data transmitted
over network, and the amount of work to be done on a coordinator.

Unfortunately, the upstream code seems not quite ready for such plans,
leading to failures like this

    ERROR:  variable not found in subplan target list

for large amounts of data and high max_parallel_workers_per_gather.

Those plans would still be quite beneficial, improving the scalability
of Postgres-XL clusters in analytics. But we can reintroduce them once
the targetlist issue gets fixed.