Ensure slab allocator is initialised correctly.
XL uses tuplesort facilities to merge incoming tuples from the datanode
connections. We'd missed out a few things during the merge and this patch
attempts to fix that.

It's not clear if we have addressed all problems in this area, since the code
has been heavily modified in PG10. But this patch at the very least allows
further testing.