Restore the regular pgbench transaction scripts
Commit c2b1e01a made script name prefixes unique by removing the regular
scripts and only keeping the XL-specific ones (with additional conditions
on bid). That turned out to be an incorrect approach - I misunderstood
the XL code and believed XL only executes the modified scripts. But those
scripts are executed only when "-k" option is specified, and the regular
ones should be used by default.

This commit fixes this by reintroducing the regular pgbench scripts, and
addresses the duplicate names by making the script names unique only when
combined with the '-k' option value.

It also adds an XL version of the 'simple-update' script with the extra
bid condition, which was missing in XL 9.5 for some reason.