Fix regression errors due to ROUNDROBIN randomization
Commit 1d14325822 randomized the initial node selection with ROUNDROBIN
distribution, but that unexpectedly broke multiple regression tests by
changing the order of returned rows (or making it unstable).

In all the tests the table ends up distributed by ROUNDROBIN implicitly
as the particular data type (e.g. "line" or "jsonb") does not support
HASH distribution.

The test data are typically inserted row-by-row, so before 1d14325822
all the rows ended up of the first datanode, which made the query order
deterministic. But since 1d14325822 the rows are distributed randomly.

Wherever possible, the order is made deterministic by adding ORDER BY
to the queries, as usual. When not possible, because the data type
does not support ordering (e.g. "line" again) the table is explicitly
defined as replication.