Accept expected plan output changes in the 'select' test case
The change was simply an addition of Remote FQS node on top the existing node.
Since the query runs an explain analyse and we don't yet support printing
analyse information underneath a Remote FQS node, also added another test to
just get explain output and ensure that we get a sane plan.